fixing some warnings
git-svn-id: https://edwdev-ik.fzk.de/svn/EDELWEISS/analysis/kdata/trunk@1368 e4a34bab-3947-0410-866c-ab59ef520b34

+++ kamping/KTrapKamperProto.h
@@ -71,10 +71,10 @@ private:
     
   double GetMean(unsigned int first, unsigned int last, double *pulse, unsigned int pulseLength, int polarity);
   
-  double RunPulseStartTime(vector<KTrapezoidalFilter *>& trapVec, KOrderFilter& ord1, KOrderFilter& ord2, 
+  unsigned int RunPulseStartTime(vector<KTrapezoidalFilter *>& trapVec, KOrderFilter& ord1, KOrderFilter& ord2, 
     KPtaProcessor& fromProcessor, int polarity);
-  double RunIonPulseStartTime(int polarity);
-  double RunHeatPulseStartTime(void);
+  unsigned int RunIonPulseStartTime(int polarity);
+  unsigned int RunHeatPulseStartTime(void);
   
   virtual KTrapezoidalFilter* AddTrapTime(vector<KTrapezoidalFilter *>& trapVec, double decay, 
                                 unsigned int rise, unsigned int flat);
